stockifai-backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
python-dotenv==1.0.1
mysqlclient==2.2.4
pandas==2.2.2
openpyxl==3.1.5
PyMySQL==1.1.1
